+++ test/harness/project/source/github_test.exs
@@ -133,7 +133,7 @@ defmodule Harness.Project.Source.GithubTest do
     %Project{
       name: name,
       source: {:github, url},
-      check_stack: %CheckStack{name: :tiny, checks: []},
+      check_stacks: [%CheckStack{name: :tiny, checks: []}],
       roadmap_path: "/tmp/#{name}"
     }
   end